weather_project/requirements.txt
# Install with: pip install -r requirements.txt
Flask==3.0.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
requests==2.32.3
python-dotenv==1.0.1
seaborn==0.13.2
matplotlib==3.9.2
